+++ data-raw/mdt-dict.R
@@ -31,6 +31,10 @@ stopifnot(!anyDuplicated(input$key),
           all(input$key == latvian$key))
 input$LV <- latvian$LV
 
+portuguese <- read.csv("data-raw/dict-portuguese.csv", stringsAsFactors = FALSE, encoding = "utf-8")
+names(portuguese)[[1]] <- "key"
+stopifnot(all(input$key == portuguese$key))
+input$PT <- portuguese$PT
 
 mdt_dict <- psychTestR::i18n_dict$new(input)
 usethis::use_data(mdt_dict, overwrite = TRUE)